feat(12): refactored JS hooks
