+++ assets/js/hooks/send_message_on_enter_key_press.js
@@ -1,0 +1,14 @@
+SendMessageOnEnterKeyPress = {
+    mounted() {
+        this.el.addEventListener("keydown", (event) => {
+            if (event.which === 13 && !event.shiftKey) {
+                textareaElement = document.getElementById("message_textarea") 
+                this.pushEvent("save_message",  {"text": textareaElement.value})
+                textareaElement.value = ""
+                textareaElement.blur()
+            } 
+        })
+    }
+}
+
+export { SendMessageOnEnterKeyPress }
