+++ assets/js/utils.js
@@ -1,0 +1,5 @@
+function extractActionAndMessageIdFromDomElementId(element) {
+    return element.id.split(/-(.*)/s)
+}
+
+export {extractActionAndMessageIdFromDomElementId}
